fixes927: the debug log has no %lu/%lx/%u -- four diagnostics were printing their format string
The fixes925/926 census produced no data at all. Every line came out as the
literal text:

  LIFE mutcensus total=%lu setattr=%lu ... CLASS=%lu STYLE=%lu

macsurf_debug_log_writef is a HAND-ROLLED formatter (macsurf_debug_log.c), not
MSL vsnprintf: it understands %d, %ld, %p, %s and %% ONLY, and emits anything
else verbatim. This is documented in CLAUDE.md and in my own notes, and I used
%lu anyway. A whole measurement round wasted.

Audited every macsurf_debug_log_writef format string in the tree and fixed all
four broken ones:
- macos9_reconvert.c  the census, %lu -> %ld with (long) casts
- talloc.c            "TALLOC CORRUPT unknown ... flags=%lx" -- the CORRUPTION
                      REPORT itself was printing "flags=%lx" instead of the
                      value, in exactly the crash class we spent this session
                      chasing. Now %ld (the formatter has no hex; the value is
                      only compared against TALLOC_MAGIC).
- macos9_deathrow.c   "STALE PIN ... passes=%u" -> %ld
- html.c              "reconvert: defer - %u fetches active" -> %ld

Remaining %u/%i/%x/%g in html.c are NSLOG (compiled out) and box_dump fprintf
(real printf), which are fine.

No behaviour change. C89-clean; harness 31/31.

+++ browser/netsurf/content/handlers/html/html.c
@@ -2320,8 +2320,8 @@ nserror html_reconvert(html_content *c)
 	 * after our free — causing a use-after-free in html_object_done.
 	 * Wait until active reaches 0 before touching the object list. */
 	if (c->base.active > 0) {
-		macsurf_debug_log_writef("reconvert: defer — %u fetches active",
-				c->base.active);
+		macsurf_debug_log_writef("reconvert: defer — %ld fetches active",
+				(long) c->base.active);
 		return NSERROR_NEED_DATA;
 	}
